+++ packages/ozone/src/api/moderation/emitEvent.ts
@@ -157,6 +157,7 @@ const handleModerationEvent = async ({
       event,
       subject,
       createdBy,
+      modTool: input.body.modTool,
     })
 
     const tagService = new TagService(
@@ -246,6 +247,7 @@ export default function (server: Server, ctx: AppContext) {
                 comment:
                   '[DIVERT_SIDE_EFFECT]: Automatically taking down after divert event',
               },
+              modTool: input.body.modTool,
             },
           },
         })