+++ packages/ozone/src/mod-service/index.ts
@@ -1,6 +1,6 @@
 import { Insertable, RawBuilder, sql } from 'kysely'
 import { CID } from 'multiformats/cid'
-import { AtpAgent } from '@atproto/api'
+import { AtpAgent, ToolsOzoneModerationDefs } from '@atproto/api'
 import { addHoursToDate, chunkArray } from '@atproto/common'
 import { Keypair } from '@atproto/crypto'
 import { IdResolver } from '@atproto/identity'
@@ -168,6 +168,7 @@ export class ModerationService {
     collections: string[]
     subjectType?: string
     policies?: string[]
+    modTool?: string[]
   }): Promise<{ cursor?: string; events: ModerationEventRow[] }> {
     const {
       subject,
@@ -189,6 +190,7 @@ export class ModerationService {
       collections,
       subjectType,
       policies,
+      modTool,
     } = opts
     const { ref } = this.db.db.dynamic
     let builder = this.db.db.selectFrom('moderation_event').selectAll()
@@ -288,6 +290,11 @@ export class ModerationService {
         })
         return qb
       })
+    }
+    if (modTool?.length) {
+      builder = builder
+        .where('modTool', 'is not', null)
+        .where(sql`("modTool" ->> 'name')`, 'in', modTool)
     }
 
     const keyset = new TimeIdKeyset(
@@ -398,12 +405,13 @@ export class ModerationService {
     subject: ModSubject
     createdBy: string
     createdAt?: Date
+    modTool?: ToolsOzoneModerationDefs.ModTool
   }): Promise<{
     event: ModerationEventRow
     subjectStatus: ModerationSubjectStatusRow | null
   }> {
     this.db.assertTransaction()
-    const { event, subject, createdBy, createdAt = new Date() } = info
+    const { event, subject, createdBy, createdAt = new Date(), modTool } = info
 
     const createLabelVals =
       isModEventLabel(event) && event.createLabelVals.length > 0
@@ -517,6 +525,7 @@ export class ModerationService {
         subjectCid: subjectInfo.subjectCid,
         subjectBlobCids: jsonb(subjectInfo.subjectBlobCids),
         subjectMessageId: subjectInfo.subjectMessageId,
+        modTool: modTool ? jsonb(modTool) : null,
       })
       .returningAll()
       .executeTakeFirstOrThrow()
@@ -811,6 +820,10 @@ export class ModerationService {
     subject: ModSubject
     reportedBy: string
     createdAt?: Date
+    modTool?: {
+      name: string
+      meta?: { [_ in string]: unknown }
+    }
   }): Promise<{
     event: ModerationEventRow
     subjectStatus: ModerationSubjectStatusRow | null
@@ -821,6 +834,7 @@ export class ModerationService {
       reportedBy,
       createdAt = new Date(),
       subject,
+      modTool,
     } = info
 
     const result = await this.logEvent({
@@ -832,6 +846,7 @@ export class ModerationService {
       createdBy: reportedBy,
       subject,
       createdAt,
+      modTool,
     })
 
     return result